Finished cleanup
Added a driver.cpp and driver.h file for organization of main and driver code. Also got rid of unnecessary headers.
