+++ over-under/include/main.h
@@ -34,7 +34,6 @@
  */
 #define PROS_USE_LITERALS
 
-#include "api.h"
 
 /**
  * You should add more #includes here
