+++ over-under/src/auton.cpp
@@ -1,14 +1,9 @@
-#include "main.h"
-#include <chrono>
 #include "global_defs.h"
-#include "pros/imu.h"
 #include "pros/imu.hpp"
-#include "pros/llemu.hpp"
-#include "pros/motors.h"
 #include "pros/motors.hpp"
 #include "auton.h"
 #include <cmath>
-#include <string>
+#include <iostream>
 #include <math.h>
 
 using namespace pros;
